Finish tetris game

+++ frontend/src/modules/tetris/InputHandler.ts
@@ -17,17 +17,19 @@ export interface UserInput {
     hold: KeyProp;
     pause: KeyProp;
     restart: KeyProp;
+    cancelRestart: KeyProp;
 }
 
 export interface MoveCallbacks {
-    moveLeft: () => void;
-    moveRight: () => void;
-    moveDown: () => void;
-    spin: () => void;
-    drop: () => void;
-    // pause: () => void;
-    // restart: () => void;
-    // hold: () => void;
+    moveLeft?: () => void;
+    moveRight?: () => void;
+    moveDown?: () => void;
+    spin?: () => void;
+    drop?: () => void;
+    pause?: () => void;
+    restart?: () => void;
+    hold?: () => void;
+    cancelRestart?: () => void;
 }
 
 export const initialInput: UserInput = {
@@ -39,6 +41,7 @@ export const initialInput: UserInput = {
     hold: { pressed: false, handleAt: 0, initialPress: true },
     pause: { pressed: false, handleAt: 0, initialPress: true },
     restart: { pressed: false, handleAt: 0, initialPress: true },
+    cancelRestart: { pressed: false, handleAt: 0, initialPress: true },
 };
 
 /**
@@ -55,9 +58,10 @@ export const setupKeyBinds = (window: Window, input: UserInput) => {
         // Handle keys only once
         if (e.key === 'ArrowUp') input.spin.pressed = true;
         if (e.key === ' ') input.drop.pressed = true;
-        if (e.key === 'z') input.hold.pressed = true;
+        if (e.key === 'c') input.hold.pressed = true;
         if (e.key === 'p') input.pause.pressed = true;
         if (e.key === 'r') input.restart.pressed = true;
+        if (e.key === 't') input.cancelRestart.pressed = true;
     };
 
     window.onkeyup = (e: KeyboardEvent) => {
@@ -67,9 +71,10 @@ export const setupKeyBinds = (window: Window, input: UserInput) => {
         // Handle keys only once
         if (e.key === 'ArrowUp') input.spin.pressed = false;
         if (e.key === ' ') input.drop.pressed = false;
-        if (e.key === 'z') input.hold.pressed = false;
+        if (e.key === 'c') input.hold.pressed = false;
         if (e.key === 'p') input.pause.pressed = false;
         if (e.key === 'r') input.restart.pressed = false;
+        if (e.key === 't') input.cancelRestart.pressed = false;
     };
 };
 
@@ -84,27 +89,50 @@ export const handleInput = (
         if (key.pressed) {
             if (frame > key.handleAt) {
                 if (key.initialPress) {
-                    if (key === input.right) moveCallbacks.moveRight();
-                    if (key === input.left) moveCallbacks.moveLeft();
-                    if (key === input.down) moveCallbacks.moveDown();
+                    if (key === input.right && moveCallbacks.moveRight) {
+                        moveCallbacks.moveRight();
+                    }
+                    if (key === input.left && moveCallbacks.moveLeft)
+                        moveCallbacks.moveLeft();
+                    if (key === input.down && moveCallbacks.moveDown)
+                        moveCallbacks.moveDown();
                     key.handleAt = frame + keyPressFrameDelay;
                     key.initialPress = false;
                 } else if (frame % framesPerKeyHandle === 0) {
-                    if (key === input.right) moveCallbacks.moveRight();
-                    if (key === input.left) moveCallbacks.moveLeft();
-                    if (key === input.down) moveCallbacks.moveDown();
+                    if (key === input.right && moveCallbacks.moveRight)
+                        moveCallbacks.moveRight();
+                    if (key === input.left && moveCallbacks.moveLeft)
+                        moveCallbacks.moveLeft();
+                    if (key === input.down && moveCallbacks.moveDown)
+                        moveCallbacks.moveDown();
                 }
             }
         } else {
             key.initialPress = true;
         }
     });
-    if (input.spin.pressed) {
+    if (input.spin.pressed && moveCallbacks.spin) {
         moveCallbacks.spin();
         input.spin.pressed = false;
     }
-    if (input.drop.pressed) {
+    if (input.drop.pressed && moveCallbacks.drop) {
         moveCallbacks.drop();
         input.drop.pressed = false;
+    }
+    if (input.hold.pressed && moveCallbacks.hold) {
+        moveCallbacks.hold();
+        input.hold.pressed = false;
+    }
+    if (input.pause.pressed && moveCallbacks.pause) {
+        moveCallbacks.pause();
+        input.pause.pressed = false;
+    }
+    if (input.restart.pressed && moveCallbacks.restart) {
+        moveCallbacks.restart();
+        input.restart.pressed = false;
+    }
+    if (input.cancelRestart.pressed && moveCallbacks.cancelRestart) {
+        moveCallbacks.cancelRestart();
+        input.cancelRestart.pressed = false;
     }
 };
